msm: kgsl: Don't make assumptions about VMA regions [LeeDroid-]
We cannot assume that a VMA region created as a result of an mmap
belongs exclusively to us.  Allow the user to pass the size of
the vmalloc region through the 'gpuaddr' member of the ioctl
struct and do strict checking on the returned vma to make sure
it is valid.

Wei Zou <wzou@codeaurora.org>
Modded by faux123